refactor: split user audience queries
